Calificar vendedor 0%->20% Calificacion.cs DateTime fecha y int puntaje convertidos a DateTime? y int? asi admiten valores null de la BD (ya que las calificaciones se crean en null, salvo el cod que se genera solo)

+++ src/FrbaCommerce/Clases/Compra.cs
@@ -48,7 +48,7 @@ namespace FrbaCommerce.Clases
             Fecha = fechaOperacion;
             if (calificacion != null)
             {
-                Calificacion = calificacion.Puntaje;
+                Calificacion = (int)calificacion.Puntaje;
                 if (calificacion.Descripcion.Equals(""))
                 {
                     Comentarios = "Sin comentarios";
